src/requirements.txt
PyYAML==6.0
tornado==6.2
